Kill off two VS warnings
git-svn-id: https://simulationcraft.googlecode.com/svn/trunk@15960 3b4652a1-8050-0410-ac47-3d40261b0f8b

+++ engine/sc_sim.cpp
@@ -1673,7 +1673,7 @@ bool progress_bar_t::update()
   status.insert( 1, steps, '.' );
   status += "]";
 
-  int length = steps * pct + 0.5;
+  int length = ( int ) ( steps * pct + 0.5 );
   for ( int i=1; i < length+1; i++ ) status[ i ] = '=';
   if ( length > 0 ) status[ length ] = '>';
